requirements.txt
allure-combine==1.0.11
allure-pytest==2.15.3
allure-python-commons==2.15.3
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
bs4==0.0.2
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
coverage==7.13.5
docopt==0.6.2
Flask==3.1.3
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpath-ng==1.8.0
MarkupSafe==3.0.3
packaging==26.0
pipreqs==0.4.13
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.1.0
pytest-html==4.2.0
pytest-metadata==3.1.1
PyYAML==6.0.3
requests==2.33.0
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.7
yarg==0.1.10

jsonpath-python
loguru